refactor(instrumentation.ts, server/): Created a new server directory to hold all non node client side scripts
Mainly to seperate startup scripts with client side stuff
